+++ scripts/run-cool2multirescool.sh
@@ -2,5 +2,5 @@
 input=$1
 ncores=$2
 outprefix=$3
-cooler zoomify $input -n $ncores -o $outprefix.multires.cool
+cooler zoomify --balance $input -n $ncores -o $outprefix.multires.cool